requirements.txt
inputs
pynput
pygame-ce
pyinstaller
